fix: don't try refresh if retriedKey exists

+++ lib/app/modules/user/data/data_sources/remote/authorize_interceptor.dart
@@ -47,14 +47,16 @@ class AuthorizeInterceptor extends Interceptor {
     RequestOptions options,
     RequestInterceptorHandler handler,
   ) async {
-    if (_repository.tokenExpiration != null) {
-      if (DateTime.now().isAfter(_repository.tokenExpiration!)) {
-        try {
-          final userApi = sl<UserApi>();
-          final token = await userApi.refresh();
-          await _repository.saveToken(token.accessToken);
-        } catch (_) {
-          await _repository.deleteToken();
+    if (!options.extra.containsKey(retriedKey)) {
+      if (_repository.tokenExpiration != null) {
+        if (DateTime.now().isAfter(_repository.tokenExpiration!)) {
+          try {
+            final userApi = sl<UserApi>();
+            final token = await userApi.refresh();
+            await _repository.saveToken(token.accessToken);
+          } catch (_) {
+            await _repository.deleteToken();
+          }
         }
       }
     }
